Coprocessor SPI integration WIP
* Get more stuff to compile on ESP32.
* SPI stuff is broken in a bad way on ESP32 for some reason. DMA memory faults?
* Refactoring.
