+++ firmware/common/CMakeLists.txt
@@ -1,10 +1,12 @@
 
 target_sources(${PROJECT_NAME} PRIVATE
-    spi_coprocessor.cc    
-    utils/buffer_utils.cc
-    utils/data_structures.cc
-    adsb/transponder_packet.cc
-    adsb/aircraft_dictionary.cc
+# NOTE: Annoyingly, all source files need to end with .c or .cpp to be seen by the ESP IDF.
+    spi_coprocessor.cpp    
+    utils/buffer_utils.cpp
+    utils/data_structures.cpp
+    adsb/transponder_packet.cpp
+    adsb/aircraft_dictionary.cpp
+    adsb/decode_utils.cpp
 )
 target_include_directories(${PROJECT_NAME} PRIVATE
     adsb
