--- firmware/common/utils/data_structures.cc
@@ -1,1 +1,0 @@
-#include "data_structures.hh"
