+++ firmware/common/utils/data_structures.cpp
@@ -1,0 +1,1 @@
+#include "data_structures.hh"
