+++ firmware/pico/host_test/comms.hh
@@ -3,8 +3,18 @@
 
 #include <cstdio>
 
-#define CONSOLE_INFO(format, ...)    printf("INFO: " format "\r\n" __VA_OPT__(, ) __VA_ARGS__);
-#define CONSOLE_WARNING(format, ...) printf("WARNING: " format "\r\n" __VA_OPT__(, ) __VA_ARGS__);
-#define CONSOLE_ERROR(format, ...)   printf("ERROR: " format "\r\n" __VA_OPT__(, ) __VA_ARGS__);
+// Use do while(0) structure to enforce semicolon usage after macro.
+#define CONSOLE_INFO(tag, format, ...)                                     \
+    do {                                                                   \
+        printf("INFO: " tag ": " format "\r\n" __VA_OPT__(, ) __VA_ARGS__) \
+    } while (0)
+#define CONSOLE_WARNING(tag, format, ...)                                     \
+    do {                                                                      \
+        printf("WARNING: " tag ": " format "\r\n" __VA_OPT__(, ) __VA_ARGS__) \
+    } while (0)
+#define CONSOLE_ERROR(tag, format, ...)                                     \
+    do {                                                                    \
+        printf("ERROR: " tag ": " format "\r\n" __VA_OPT__(, ) __VA_ARGS__) \
+    } while (0)
 
 #endif /* COMMS_HH_ */
