+++ firmware/pico/src/app/ads_bee.cc
@@ -86,7 +86,7 @@ bool ADSBee::Init() {
     gpio_set_function(config_.onboard_i2c_scl_pin, GPIO_FUNC_I2C);
     uint8_t wiper_value_counts;
     if (i2c_read_blocking(config_.onboard_i2c, kRxGainDigipotI2CAddr, &wiper_value_counts, 1, false) != 1) {
-        CONSOLE_ERROR("ADSBee::Init: Failed to read wiper position from Rx Gain Digipot at I2C address 0x%x.",
+        CONSOLE_ERROR("ADSBee::Init", "Failed to read wiper position from Rx Gain Digipot at I2C address 0x%x.",
                       kRxGainDigipotI2CAddr);
         return false;
     }
@@ -131,7 +131,7 @@ bool ADSBee::Init() {
                                      message_demodulator_offset_, config_.pulses_pin, config_.recovered_clk_pin,
                                      message_demodulator_div);
 
-    CONSOLE_INFO("ADSBee::Init: PIOs initialized.");
+    CONSOLE_INFO("ADSBee::Init", "PIOs initialized.");
 
     gpio_init(config_.status_led_pin);
     gpio_set_dir(config_.status_led_pin, GPIO_OUT);
@@ -253,8 +253,8 @@ uint64_t ADSBee::GetMLAT12MHzCounts(uint16_t num_bits) {
 
 bool ADSBee::SetTLHiMilliVolts(int tl_hi_mv) {
     if (tl_hi_mv > kTLMaxMV || tl_hi_mv < kTLMinMV) {
-        CONSOLE_PRINTF(
-            "ADSBee::SetTLHiMilliVolts: Unable to set tl_hi_mv_ to %d, outside of permissible range %d-%d.\r\n",
+        CONSOLE_ERROR(
+            "ADSBee::SetTLHiMilliVolts", "Unable to set tl_hi_mv_ to %d, outside of permissible range %d-%d.\r\n",
             tl_hi_mv, kTLMinMV, kTLMaxMV);
         return false;
     }
@@ -266,8 +266,8 @@ bool ADSBee::SetTLHiMilliVolts(int tl_hi_mv) {
 
 bool ADSBee::SetTLLoMilliVolts(int tl_lo_mv) {
     if (tl_lo_mv > kTLMaxMV || tl_lo_mv < kTLMinMV) {
-        CONSOLE_PRINTF(
-            "ADSBee::SetTLLoMilliVolts: Unable to set tl_lo_mv_ to %d, outside of permissible range %d-%d.\r\n",
+        CONSOLE_ERROR(
+            "ADSBee::SetTLLoMilliVolts", "Unable to set tl_lo_mv_ to %d, outside of permissible range %d-%d.\r\n",
             tl_lo_mv, kTLMinMV, kTLMaxMV);
         return false;
     }
