+++ firmware/pico/src/app/settings.cc
@@ -16,7 +16,7 @@ const char SettingsManager::ReportingProtocolStrs[SettingsManager::ReportingProt
 
 bool SettingsManager::Load() {
     if (!eeprom.Load(settings)) {
-        CONSOLE_ERROR("settings.cc::Load: Failed load settings from EEPROM.");
+        CONSOLE_ERROR("settings.cc::Load", "Failed load settings from EEPROM.");
         return false;
     };
 
@@ -24,7 +24,7 @@ bool SettingsManager::Load() {
     if (settings.magic_word != kSettingsVersionMagicWord) {
         ResetToDefaults();
         if (!eeprom.Save(settings)) {
-            CONSOLE_ERROR("settings.cc::Load: Failed to save default settings.");
+            CONSOLE_ERROR("settings.cc::Load", "Failed to save default settings.");
             return false;
         }
     }
