--- firmware/pico/src/hal/hal.hh
@@ -1,9 +1,0 @@
-#ifndef HAL_HH_
-#define HAL_HH_
-
-#include "stdint.h"
-
-uint64_t get_time_since_boot_us();
-uint32_t get_time_since_boot_ms();
-
-#endif /* HAL_HH_ */
